requirements.txt
cachetools
svgwrite
falcon
requests
PyJWT
cryptography
